Solving Issue/Bug 4.
By mistake, I changed score label with health label, and so when the
enemy hits the player, the score decreases instead of the health, and
when the player hits an enemy, the health increases instead of the
score. All I need to change is that I write score label with score
counter, and health label with health counter.