Solving Issue/Bug 5
When the player loses the game, the game will still display the game won
image instead of the game over image. I checked my code and found out
that by mistake I called the game won image. All I need to do is to
change Game Won written in the code with game over in the Game Over
javascript.